requirements.txt
numpy<=2.1.1
flask==3.0.3
opencv-python==4.11.0.86
ultralytics==8.3.89
python-can==4.5.0
roboflow==1.1.58